Give python package a name
Package name change from `signature` to `src` caused a few issues, including breaking test.py and preventing normal dotted name imports.  If we want a `/src` subfolder, we should still have a valid subfolder with a package name.  Typically, the package name is the same as the repository name, but we can choose something better if we want.